+++ OpenPOS-APP/MauiProgram.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using OpenPOS_APP.Models;
 using OpenPOS_APP.Services;
+using OpenPOS_APP.Services.Models;
 
 namespace OpenPOS_APP;
