+++ OpenPOS-APP/Services/Models/OrderService.cs
@@ -1,31 +1,31 @@
 using OpenPOS_APP.Models;
 using OpenPOS_APP.Services.Interfaces;
 
-namespace OpenPOS_APP.Services;
+namespace OpenPOS_APP.Services.Models;
 
-public class FloorService: IModelService<Floor>
+public class OrderService : IModelService<Order>
 {
-    public static List<Floor> GetAll()
+    public static List<Order> GetAll()
     {
         throw new NotImplementedException();
     }
 
-    public static Floor FindByID(int id)
+    public static Order FindByID(int id)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Delete(Floor obj)
+    public static bool Delete(Order obj)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Update(Floor obj)
+    public static bool Update(Order obj)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Create(Floor obj)
+    public static bool Create(Order obj)
     {
         throw new NotImplementedException();
     }
