+++ OpenPOS-APP/Services/Models/RoleService.cs
@@ -1,31 +1,31 @@
 using OpenPOS_APP.Models;
 using OpenPOS_APP.Services.Interfaces;
 
-namespace OpenPOS_APP.Services;
+namespace OpenPOS_APP.Services.Models;
 
-public class CategoryService: IModelService<Category>
+public class RoleService : IModelService<Role>
 {
-    public static List<Category> GetAll()
+    public static List<Role> GetAll()
     {
         throw new NotImplementedException();
     }
 
-    public static Category FindByID(int id)
+    public static Role FindByID(int id)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Delete(Category obj)
+    public static bool Delete(Role obj)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Update(Category obj)
+    public static bool Update(Role obj)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Create(Category obj)
+    public static bool Create(Role obj)
     {
         throw new NotImplementedException();
     }
