+++ OpenPOS-APP/Services/Models/TableService.cs
@@ -1,31 +1,31 @@
 using OpenPOS_APP.Models;
 using OpenPOS_APP.Services.Interfaces;
 
-namespace OpenPOS_APP.Services;
+namespace OpenPOS_APP.Services.Models;
 
-public class RoleService : IModelService<Role>
+public class TableService : IModelService<Table>
 {
-    public static List<Role> GetAll()
+    public static List<Table> GetAll()
     {
         throw new NotImplementedException();
     }
 
-    public static Role FindByID(int id)
+    public static Table FindByID(int id)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Delete(Role obj)
+    public static bool Delete(Table obj)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Update(Role obj)
+    public static bool Update(Table obj)
     {
         throw new NotImplementedException();
     }
 
-    public static bool Create(Role obj)
+    public static bool Create(Table obj)
     {
         throw new NotImplementedException();
     }
